Revert "Renamed to GlobalLink Now"
This reverts commit c5f5583dd3fdcec87a91e30e7f2a31ee3160fad6.
