+++ Apps.GlobalLinkAI/Api/AppRequest.cs
@@ -1,0 +1,19 @@
+using Apps.GlobalLinkAI.Constants;
+using Blackbird.Applications.Sdk.Common.Authentication;
+using Blackbird.Applications.Sdk.Utils.Extensions.Sdk;
+using Blackbird.Applications.Sdk.Utils.RestSharp;
+using RestSharp;
+
+namespace Apps.GlobalLinkAI.Api;
+
+public class AppRequest : BlackBirdRestRequest
+{
+    public AppRequest(string resource, Method method, IEnumerable<AuthenticationCredentialsProvider> creds) : base(resource, method, creds)
+    {
+    }
+
+    protected override void AddAuth(IEnumerable<AuthenticationCredentialsProvider> creds)
+    {
+        this.AddHeader("subscription-key", creds.Get(CredsNames.ApiKey).Value);
+    }
+}
