+++ Apps.GlobalLinkAI/Connections/ConnectionDefinition.cs
@@ -1,0 +1,29 @@
+﻿using Apps.GlobalLinkAI.Constants;
+using Blackbird.Applications.Sdk.Common.Authentication;
+using Blackbird.Applications.Sdk.Common.Connections;
+
+namespace Apps.GlobalLinkAI.Connections;
+
+public class ConnectionDefinition : IConnectionDefinition
+{
+    public IEnumerable<ConnectionPropertyGroup> ConnectionPropertyGroups => new List<ConnectionPropertyGroup>
+    {
+        new()
+        {
+            Name = "Developer API key",
+            AuthenticationType = ConnectionAuthenticationType.Undefined,
+            ConnectionUsage = ConnectionUsage.Actions,
+            ConnectionProperties = new List<ConnectionProperty>
+            {
+                new(CredsNames.Host) { DisplayName = "Host" },
+                new(CredsNames.ApiKey) { DisplayName = "API key", Sensitive = true }
+            }
+        }
+    };
+
+    public IEnumerable<AuthenticationCredentialsProvider> CreateAuthorizationCredentialsProviders(
+        Dictionary<string, string> values) =>
+        values.Select(x =>
+                new AuthenticationCredentialsProvider(AuthenticationCredentialsRequestLocation.None, x.Key, x.Value))
+            .ToList();
+}
