+++ Apps.GlobalLinkAI/Constants/CredsNames.cs
@@ -1,0 +1,7 @@
+namespace Apps.GlobalLinkAI.Constants;
+
+public static class CredsNames
+{
+    public const string ApiKey = "api_key";
+    public const string Host = "host";
+}
