+++ Apps.GlobalLinkAI/DataSourceHandlers/StaticDataSourceHandlers/TextTypeDataHandler.cs
@@ -1,0 +1,12 @@
+using Blackbird.Applications.Sdk.Utils.Sdk.DataSourceHandlers;
+
+namespace Apps.GlobalLinkAI.DataSourceHandlers.StaticDataSourceHandlers;
+
+public class TextTypeDataHandler : EnumDataHandler
+{
+    protected override Dictionary<string, string> EnumValues => new()
+    {
+        ["text"] = "Text",
+        ["html"] = "HTML",
+    };
+}
