+++ Apps.GlobalLinkAI/Models/Entities/EngineEntity.cs
@@ -1,0 +1,10 @@
+namespace Apps.GlobalLinkAI.Models.Entities;
+
+public class EngineEntity
+{
+    public string Id { get; set; }
+    
+    public string Alias { get; set; }
+    
+    public IEnumerable<string> SupportedLanguages { get; set; }
+}
