+++ Apps.GlobalLinkAI/Models/Entities/TextTranslationEntity.cs
@@ -1,0 +1,6 @@
+namespace Apps.GlobalLinkAI.Models.Entities;
+
+public class TextTranslationEntity
+{
+    public string Text { get; set; }
+}
