+++ Apps.GlobalLinkAI/Models/Request/FileRequest.cs
@@ -1,0 +1,8 @@
+using Blackbird.Applications.Sdk.Common.Files;
+
+namespace Apps.GlobalLinkAI.Models.Request;
+
+public class FileRequest
+{
+    public FileReference File { get; set; }
+}
