+++ Apps.GlobalLinkAI/Models/Response/Engines/ListEnginesResponse.cs
@@ -1,0 +1,8 @@
+using Apps.GlobalLinkAI.Models.Entities;
+
+namespace Apps.GlobalLinkAI.Models.Response.Engines;
+
+public class ListEnginesResponse
+{
+    public IEnumerable<EngineEntity> Engines { get; set; }
+}
