+++ Apps.GlobalLinkAI/Models/Response/ErrorResponse.cs
@@ -1,0 +1,6 @@
+namespace Apps.GlobalLinkAI.Models.Response;
+
+public class ErrorResponse
+{
+    public string Message { get; set; }
+}
