+++ Apps.GlobalLinkAI/Models/Response/FileResponse.cs
@@ -1,0 +1,8 @@
+using Blackbird.Applications.Sdk.Common.Files;
+
+namespace Apps.GlobalLinkAI.Models.Response;
+
+public class FileResponse
+{
+    public FileReference File { get; set; }
+}
