+++ Apps.GlobalLinkAI/Models/Response/Translation/TextTranslationResponse.cs
@@ -1,0 +1,6 @@
+namespace Apps.GlobalLinkAI.Models.Response.Translation;
+
+public class TextTranslationResponse
+{
+    public IEnumerable<TranslationMessageResponse> Message { get; set; }
+}
