+++ Apps.GlobalLinkAI/Models/Response/Translation/TranslateDocumentResponse.cs
@@ -1,0 +1,6 @@
+namespace Apps.GlobalLinkAI.Models.Response.Translation;
+
+public class TranslateDocumentResponse
+{
+    public IEnumerable<TranslateFileResponse> FileIds { get; set; }
+}
