+++ Apps.GlobalLinkAI/Models/Response/Translation/TranslateFileResponse.cs
@@ -1,0 +1,10 @@
+namespace Apps.GlobalLinkAI.Models.Response.Translation;
+
+public class TranslateFileResponse
+{
+    public string FileId { get; set; }
+    
+    public string Status { get; set; }
+    
+    public string ErrorCode { get; set; }
+}
