+++ Apps.GlobalLinkAI/Models/Response/Translation/TranslationMessageResponse.cs
@@ -1,0 +1,6 @@
+namespace Apps.GlobalLinkAI.Models.Response.Translation;
+
+public class TranslationMessageResponse
+{
+    public string Text { get; set; }
+}
