--- Apps.GlobalLinkNow/Apps.GlobalLinkNow.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <Product>GlobalLink Now</Product>
-        <Description>AI-Powered Machine Translation</Description>
-        <Version>1.0.0</Version>
-        <AssemblyName>Apps.GlobalLinkNow</AssemblyName>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Blackbird.Applications.Sdk.Utils" Version="1.0.25" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <EmbeddedResource CopyToOutputDirectory="Always" Include="image\icon.png"/>
-    </ItemGroup>
-
-</Project>
