--- Apps.GlobalLinkNow/Constants/CredsNames.cs
@@ -1,7 +1,0 @@
-namespace Apps.GlobalLinkNow.Constants;
-
-public static class CredsNames
-{
-    public const string ApiKey = "api_key";
-    public const string Host = "host";
-}
