--- Apps.GlobalLinkNow/Invocables/AppInvocable.cs
@@ -1,18 +1,0 @@
-using Apps.GlobalLinkNow.Api;
-using Blackbird.Applications.Sdk.Common;
-using Blackbird.Applications.Sdk.Common.Authentication;
-using Blackbird.Applications.Sdk.Common.Invocation;
-
-namespace Apps.GlobalLinkNow.Invocables;
-
-public class AppInvocable : BaseInvocable
-{
-    protected AuthenticationCredentialsProvider[] Creds =>
-        InvocationContext.AuthenticationCredentialsProviders.ToArray();
-
-    protected AppClient Client { get; }
-    public AppInvocable(InvocationContext invocationContext) : base(invocationContext)
-    {
-        Client = new(Creds);
-    }
-}
