--- Apps.GlobalLinkNow/Models/Entities/EngineEntity.cs
@@ -1,10 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Entities;
-
-public class EngineEntity
-{
-    public string Id { get; set; }
-    
-    public string Alias { get; set; }
-    
-    public IEnumerable<string> SupportedLanguages { get; set; }
-}
