--- Apps.GlobalLinkNow/Models/Entities/TextTranslationEntity.cs
@@ -1,6 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Entities;
-
-public class TextTranslationEntity
-{
-    public string Text { get; set; }
-}
