--- Apps.GlobalLinkNow/Models/Request/FileRequest.cs
@@ -1,8 +1,0 @@
-using Blackbird.Applications.Sdk.Common.Files;
-
-namespace Apps.GlobalLinkNow.Models.Request;
-
-public class FileRequest
-{
-    public FileReference File { get; set; }
-}
