--- Apps.GlobalLinkNow/Models/Response/Engines/ListEnginesResponse.cs
@@ -1,8 +1,0 @@
-using Apps.GlobalLinkNow.Models.Entities;
-
-namespace Apps.GlobalLinkNow.Models.Response.Engines;
-
-public class ListEnginesResponse
-{
-    public IEnumerable<EngineEntity> Engines { get; set; }
-}
