--- Apps.GlobalLinkNow/Models/Response/ErrorResponse.cs
@@ -1,6 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Response;
-
-public class ErrorResponse
-{
-    public string Message { get; set; }
-}
