--- Apps.GlobalLinkNow/Models/Response/FileResponse.cs
@@ -1,8 +1,0 @@
-using Blackbird.Applications.Sdk.Common.Files;
-
-namespace Apps.GlobalLinkNow.Models.Response;
-
-public class FileResponse
-{
-    public FileReference File { get; set; }
-}
