--- Apps.GlobalLinkNow/Models/Response/Translation/TextTranslationResponse.cs
@@ -1,6 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Response.Translation;
-
-public class TextTranslationResponse
-{
-    public IEnumerable<TranslationMessageResponse> Message { get; set; }
-}
