--- Apps.GlobalLinkNow/Models/Response/Translation/TranslateDocumentResponse.cs
@@ -1,6 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Response.Translation;
-
-public class TranslateDocumentResponse
-{
-    public IEnumerable<TranslateFileResponse> FileIds { get; set; }
-}
