--- Apps.GlobalLinkNow/Models/Response/Translation/TranslateFileResponse.cs
@@ -1,10 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Response.Translation;
-
-public class TranslateFileResponse
-{
-    public string FileId { get; set; }
-    
-    public string Status { get; set; }
-    
-    public string ErrorCode { get; set; }
-}
