--- Apps.GlobalLinkNow/Models/Response/Translation/TranslationMessageResponse.cs
@@ -1,6 +1,0 @@
-namespace Apps.GlobalLinkNow.Models.Response.Translation;
-
-public class TranslationMessageResponse
-{
-    public string Text { get; set; }
-}
